only enable label editing when a node is selected
Previously, the labeling was enabled every time the labels checkbox was checked. This was a little confusing, since it allowed for the text to be edited, even if no nodes were selected, which in result did nothing.

This way, it's way easier to see, what is being changed  by the user typing stuff into the labels line edit.